client: categorize events
