+++ client/js/events/client.js
@@ -6,15 +6,14 @@ Event.register(alt.Enums.EventType.KEYBOARD_EVENT, "KeyBoardEvent");
 Event.register(alt.Enums.EventType.WEB_VIEW_EVENT, "WebViewEvent");
 Event.register(alt.Enums.EventType.WEB_SOCKET_CLIENT_EVENT, "WebSocketEvent");
 Event.register(alt.Enums.EventType.AUDIO_EVENT, "AudioEvent");
-Event.register(alt.Enums.EventType.TASK_CHANGE, "TaskChange");
 Event.register(alt.Enums.EventType.SPAWNED, "Spawned");
 Event.register(alt.Enums.EventType.RMLUI_EVENT, "RmluiEvent");
 Event.register(alt.Enums.EventType.WINDOW_FOCUS_CHANGE, "WindowFocusChange");
 Event.register(alt.Enums.EventType.WINDOW_RESOLUTION_CHANGE, "WindowResolutionChange");
 
 // Custom keyboard events
-Event.register(alt.Enums.CustomEventType.KEY_UP, "Keyup", true);
-Event.register(alt.Enums.CustomEventType.KEY_DOWN, "Keydown", true);
+Event.register(alt.Enums.CustomEventType.KEY_UP, "KeyUp", true);
+Event.register(alt.Enums.CustomEventType.KEY_DOWN, "KeyDown", true);
 
 alt.Events.onKeyBoardEvent(({ state, key }) => {
     if (state == "UP") Event.invoke(alt.Enums.CustomEventType.KEY_UP, { key: key }, true);
