+++ client/src/events/ClientEvent.cpp
@@ -1,4 +1,5 @@
 #include "Event.h"
+#include "events/CAudioEvent.h"
 
 // clang-format off
 static js::Event keyBoardEvent(alt::CEvent::Type::KEYBOARD_EVENT, [](const alt::CEvent* ev, js::Event::EventArgs& args)
@@ -27,16 +28,9 @@ static js::Event webSocketClientEvent(alt::CEvent::Type::WEB_SOCKET_CLIENT_EVENT
 static js::Event audioEvent(alt::CEvent::Type::AUDIO_EVENT, [](const alt::CEvent* ev, js::Event::EventArgs& args)
 {
     auto e = static_cast<const alt::CAudioEvent*>(ev);
-    args.Set("target", e->GetElement());
+    args.Set("target", e->GetTarget());
     args.Set("name", e->GetName());
     args.Set("args", e->GetArgs());
-});
-
-static js::Event taskChangeEvent(alt::CEvent::Type::TASK_CHANGE, [](const alt::CEvent* ev, js::Event::EventArgs& args)
-{
-    auto e = static_cast<const alt::CTaskChangeEvent*>(ev);
-    args.Set("oldTask", e->GetOldTask());
-    args.Set("newTask", e->GetNewTask());
 });
 
 static js::Event rmluiEvent(alt::CEvent::Type::RMLUI_EVENT, [](const alt::CEvent* ev, js::Event::EventArgs& args)
